backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
chromadb
pydantic
python-dotenv
pypdf
markdown
pymongo
ollama
pdfminer.six
python-multipart